Extract common modules to separated 'common' library module
