+++ features/browser/applicationServices/src/main/java/com/diskin/alon/appsbrowser/browser/applicationservices/MyClass.java
@@ -1,0 +1,4 @@
+package com.diskin.alon.appsbrowser.browser.applicationservices;
+
+public class MyClass {
+}
