+++ features/browser/presentation/src/androidTest/java/com/diskin/alon/appsbrowser/browser/di/TestAppModule.java
@@ -1,7 +1,7 @@
 package com.diskin.alon.appsbrowser.browser.di;
 
-import com.diskin.alon.appsbrowser.browser.BrowserFragment;
-import com.diskin.alon.appsbrowser.browser.BrowserViewModel;
+import com.diskin.alon.appsbrowser.browser.controller.BrowserFragment;
+import com.diskin.alon.appsbrowser.browser.viewmodel.BrowserViewModel;
 
 import org.mockito.Mockito;
