+++ features/browser/presentation/src/main/java/com/diskin/alon/appsbrowser/browser/model/UserApp.java
@@ -1,4 +1,4 @@
-package com.diskin.alon.appsbrowser.browser;
+package com.diskin.alon.appsbrowser.browser.model;
 
 import java.util.Objects;
