+++ features/common/src/main/java/com/diskin/alon/appsbrowser/common/Mapper.java
@@ -1,0 +1,6 @@
+package com.diskin.alon.appsbrowser.common;
+
+public interface Mapper<S,R> {
+
+    R map(S source);
+}
