+++ features/common/src/main/java/com/diskin/alon/appsbrowser/common/UseCase.java
@@ -1,0 +1,4 @@
+package com.diskin.alon.appsbrowser.common;
+
+public interface UseCase<P,R> {
+}
