+++ features/common/src/test/java/com/diskin/alon/appsbrowser/common/UseCaseMediatorTest.java
@@ -1,0 +1,8 @@
+package com.diskin.alon.appsbrowser.common;
+
+/**
+ * {@link UseCaseMediator} unit test class.
+ */
+public class UseCaseMediatorTest {
+
+}
