+++ settings.gradle
@@ -1,4 +1,5 @@
 include ':app',
+        ':features:browser:applicationServices',
         ':features:home',
         ':features:browser:presentation',
         ':features:common',
